# connectors/docs/swagger.yaml
definitions:
...
  fiber.Map:
    additionalProperties: {}
    type: object
...
  irminmodels.ConnectionOAuthConfig:
    properties:
      authorization_url:
        description: |-
          AuthorizationURL is the user-agent-facing endpoint where the end
          user approves the connection. Absolute https:// URL.
        example: https://app.hubspot.com/oauth/authorize
        type: string
      dcr_endpoint:
        description: |-
          DCREndpoint is optional (RFC 7591). When set, the Core can
          dynamically register an OAuth client per workspace on first use,
          skipping the need for an admin-configured app. MCP-compatible
          vendors (Intercom, Linear, Sentry, ...) typically provide this.
        example: https://api.example.com/oauth/register
        type: string
      extra_params:
        additionalProperties:
          type: string
        description: |-
          ExtraParams are vendor-specific parameters appended to the
          authorization request (e.g., access_type=offline, prompt=consent).
          Keys that would override security-critical OAuth params are
          silently dropped by the Core.
        example:
          access_type: offline
        type: object
      pkce:
        description: |-
          PKCE must be true for new connectors. The Core refuses to run a
          flow when this is false. Retained as a field (rather than an
          always-true default) so the contract is explicit on the wire.
        example: true
        type: boolean
      provider:
        description: |-
          Provider is a short canonical identifier for the vendor, e.g.
          "hubspot", "stripe", "intercom". Used in logs + error messages.
        example: hubspot
        maxLength: 64
        type: string
      revocation_url:
        description: |-
          RevocationURL is optional (RFC 7009). When set, the Core POSTs the
          token here on disconnect so the vendor drops the grant immediately.
        example: https://api.example.com/oauth/revoke
        type: string
      scopes:
        description: |-
          Scopes is the set of OAuth scopes the connector declares it needs.
          Read-only by default; connectors should follow least-privilege.
        example:
        - crm.objects.contacts.read
        - crm.objects.deals.read
        items:
          type: string
        type: array
      token_url:
        description: TokenURL is the token endpoint (RFC 6749 §3.2). Absolute https://
          URL.
        example: https://api.hubapi.com/oauth/v1/token
        type: string
      userinfo_url:
        description: |-
          UserinfoURL is optional. When set, the Core may call it after a
          successful exchange to populate a friendly "connected as <who>"
          label in the UI. Not used in the Phase 2 wiring; reserved.
        example: https://api.example.com/oauth/userinfo
        type: string
    required:
    - authorization_url
    - provider
    - scopes
    - token_url
    type: object
  irminmodels.ConnectorCapability:
    enum:
    - pull
    - push
    - apply_patch
    - patch_event
    type: string
    x-enum-varnames:
    - ConnectorCapabilityPull
    - ConnectorCapabilityPush
    - ConnectorCapabilityApplyPatch
    - ConnectorCapabilityPatchEvent
  irminmodels.ConnectorCategory:
    enum:
    - database
    - crm
    - erp
    - warehouse
    - marketing
    - analytics
    - storage
    - messaging
    - payment
    - social
    - calendar
    - project_management
    - ecommerce
    - iot
    - monitoring
    - other
    type: string
    x-enum-varnames:
    - ConnectorCategoryDatabase
    - ConnectorCategoryCRM
    - ConnectorCategoryERP
    - ConnectorCategoryWarehouse
    - ConnectorCategoryMarketing
    - ConnectorCategoryAnalytics
    - ConnectorCategoryStorage
    - ConnectorCategoryMessaging
    - ConnectorCategoryPayment
    - ConnectorCategorySocial
    - ConnectorCategoryCalendar
    - ConnectorCategoryProjectManagement
    - ConnectorCategoryECommerce
    - ConnectorCategoryIoT
    - ConnectorCategoryMonitoring
    - ConnectorCategoryOther
...
  irminmodels.DynamicField:
    properties:
      default:
        example: Default value
        type: string
      example:
        example: Example value
        maxLength: 100
        type: string
      help_text:
        example: This is a help text
        maxLength: 200
        type: string
      label:
        example: Name
        maxLength: 100
        type: string
      max:
        example: 100
        type: number
      min:
        example: 1
        type: number
      multiple:
        example: false
        type: boolean
      options:
        items:
          $ref: '#/definitions/irminmodels.SelectOption'
        type: array
      required:
        example: true
        type: boolean
      required_with:
        example:
        - field1
        - field2
        items:
          type: string
        type: array
      secret:
        example: false
        type: boolean
      type:
        allOf:
        - $ref: '#/definitions/irminmodels.FieldType'
        enum:
        - text
        - textarea
        - password
        - email
        - checkbox
        - integer
        - float
        - select
        - radio
        - file
        - date
        - time
        - datetime
        example: text
    required:
    - label
    - type
    type: object
  irminmodels.FieldType:
    enum:
    - text
    - textarea
    - password
    - email
    - checkbox
    - integer
    - float
    - select
    - radio
    - file
    - date
    - time
    - datetime
    type: string
    x-enum-varnames:
    - FieldTypeText
    - FieldTypeTextarea
    - FieldTypePassword
    - FieldTypeEmail
    - FieldTypeCheckbox
    - FieldTypeInteger
    - FieldTypeFloat
    - FieldTypeSelect
    - FieldTypeRadio
    - FieldTypeFile
    - FieldTypeDate
    - FieldTypeTime
    - FieldTypeDatetime
...
  irminmodels.SelectOption:
    properties:
      key:
        example: key1
        maxLength: 50
        type: string
      value:
        example: Value 1
        maxLength: 100
        type: string
    required:
    - key
    - value
    type: object
...
  models.ConnectorDetails:
    properties:
      api_base_url:
        example: https://connectors.irmin.co/mysql
        type: string
      author:
        example: Irmin Team
        type: string
      author_email:
        example: hello@irmin.co
        type: string
      capabilities:
        example:
        - pull
        - push
        items:
          $ref: '#/definitions/irminmodels.ConnectorCapability'
          enum:
          - pull
          - push
          - apply_patch
          - patch_event
        type: array
      categories:
        example:
        - database
        - crm
        - erp
        - warehouse
        - marketing
        - analytics
        - storage
        - messaging
        - payment
        - social
        - calendar
        - project_management
        - ecommerce
        - iot
        - monitoring
        - other
        items:
          $ref: '#/definitions/irminmodels.ConnectorCategory'
        type: array
      connection_oauth_config:
        allOf:
        - $ref: '#/definitions/irminmodels.ConnectionOAuthConfig'
        description: |-
          ConnectionOAuthConfig is optional. When set, the connector declares
          it uses OAuth 2.0 (authorization code + PKCE) for authenticating a
          Connection, and Irmin Core runs the flow on the user's behalf. Nil
          means the connector uses the legacy DynamicField form path
          (password / API key / etc.). See OAUTH-CONNECTORS.md for the flow
          + patterns + which vendors are planned.
      description:
        example: Connect to MySQL databases for data extraction and synchronization
        type: string
      logo_url:
        example: https://cdn.irmin.co/connectors/mysql.png
        type: string
      name:
        example: MySQL Database Connector
        type: string
      primary_category:
        allOf:
        - $ref: '#/definitions/irminmodels.ConnectorCategory'
        example: database
      read_more_url:
        example: https://docs.irmin.co/connectors/mysql
        type: string
      structure_version:
        example: 1.0.0
        type: string
      version:
        example: 1.2.3
        type: string
    type: object
host: connectors.irmin.co
info:
  contact:
    email: hello@irmin.co
    name: Irmin Team
    url: https://irmin.co
  description: Collection of connectors for Irmin to interact with external services
  termsOfService: http://irmin.co/terms
  title: Irmin Connectors
  version: 1.0.0
paths:
...
  /googledrive/configuration/{key}/fields:
    post:
      consumes:
      - application/json
      - multipart/form-data
      description: Returns dynamic configuration field definitions for the Google
        Drive connector. The "details" key returns an empty map because Google Drive
        authenticates via OAuth; the "settings" key returns user-tunable options like
        scope and max_records_per_resource.
      parameters:
      - description: Configuration key
        enum:
        - details
        - settings
        in: path
        name: key
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Configuration fields
          schema:
            additionalProperties:
              $ref: '#/definitions/irminmodels.DynamicField'
            type: object
        "400":
          description: Bad request
          schema:
            $ref: '#/definitions/fiber.Map'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/fiber.Map'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/fiber.Map'
      security:
      - SystemTokenAuth: []
      summary: Get Google Drive connector configuration fields
      tags:
      - googledrive
...
  /http/operation/push:
    post:
      consumes:
      - multipart/form-data
      description: Send file content to the configured HTTP endpoint using the specified
        method and headers. Use the path parameter to modify the request URL (absolute
        path replaces, relative path appends).
      parameters:
      - description: File to send to HTTP endpoint
        in: formData
        name: file
        required: true
        type: file
      - description: Path to append/replace in the request URL (e.g., /api/users or
          customers)
        in: formData
        name: path
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Data pushed successfully
          schema:
            $ref: '#/definitions/fiber.Map'
        "400":
          description: Bad request - invalid operation token or file
          schema:
            $ref: '#/definitions/fiber.Map'
        "401":
          description: Unauthorized - invalid or missing authentication
          schema:
            $ref: '#/definitions/fiber.Map'
        "404":
          description: Operation not found
          schema:
            $ref: '#/definitions/fiber.Map'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/fiber.Map'
      security:
      - SystemTokenAuth: []
      summary: Push data to HTTP endpoint
      tags:
      - http
...
  /mysql/info:
    get:
      consumes:
      - application/json
      description: Get detailed information about the MySQL connector including capabilities,
        configuration fields, and API endpoints
      produces:
      - application/json
      responses:
        "200":
          description: MySQL connector information retrieved successfully
          schema:
            $ref: '#/definitions/models.ConnectorDetails'
        "401":
          description: Unauthorized - invalid or missing authentication
          schema:
            $ref: '#/definitions/fiber.Map'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/fiber.Map'
      security:
      - SystemTokenAuth: []
      summary: Get MySQL connector information
      tags:
      - mysql
...
  /pinecone/info:
    get:
      consumes:
      - application/json
      description: Get detailed information about the Pinecone connector including
        capabilities, configuration fields, and API endpoints
      produces:
      - application/json
      responses:
        "200":
          description: Pinecone connector information retrieved successfully
          schema:
            $ref: '#/definitions/models.ConnectorDetails'
        "401":
          description: Unauthorized - invalid or missing authentication
          schema:
            $ref: '#/definitions/fiber.Map'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/fiber.Map'
      security:
      - SystemTokenAuth: []
      summary: Get Pinecone connector information
      tags:
      - pinecone
...
schemes:
- https
securityDefinitions:
  OperationTokenAuth:
    description: Operation token with Bearer prefix - received from operation/init
      and used for data operation endpoints
    in: header
    name: Authorization
    type: apiKey
  SystemTokenAuth:
    description: System token with Bearer prefix - used for connector registration
      and initialization endpoints
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"